Fetch file data from a peer!
Only supports one peer, but does support files larger than
one block.

Also, don't copy the FileInfo when returning it. Those blocks
aren't going to be modified, so there isn't a danger of some
stupid thread interplay. (There is, of course, some danger of
anyone getting a FileInfo modifying it. But we won't do that.)