Pull file blocks in parallel from peers
All blocks are requested at once, for now. So, if
you're asking for all of a 1GB file, a bunch of
network requests are going to go out at once.

For a given block, the device to pull from is randomized.
ST prime does some game with latency and history of
requests to a device. I didn't want to take the time to
implement that at this moment, but is a possible future
improvement.