Pin files for offline availability (Github Issue #4)
The UI provides the ability to pin and unpin files. If
pins are changed, SyncthingFUSE needs to be restarted.
If files are pinned, the UI shows some progress on any
pending files, or a little info on the pinned files.

On startup and receiving updates from peers, any missing
blocks for pinned files are added to a list. Background
workers (currently hardcoded to 4) pull those blocks from
peers. If the data is needed for a read now, then the
normal read routine happens, then the background worker
is simply required for pinning the block.

The model is responsible for lots of coordination, and the
block cache is responsible storing the actual data.

The block cache performs some fancy footwork around pinned
files. Pinned blocks are still tracked in the cache access
pattern. So, if a file is un-pinned, the cache will act as
if the file was never pinned, relying on previous access
patterns for pinning. But, pinned files don't count toward
cache usage. This is managed by counting the blocks as
size 0 in the cache tracking.

On the autocomplete, the first pass just provides a very
basic autocomplete for simplicity. You have to know a
little about what you're looking for, and it only looks
one level down.

The locking in the model is a bit more complicated, now.
Sorry about that. There is an order that locks are
acquired, and hopefully it's followed everywhere. Just
wait until we decide to do per-folder locking :/

Move finding connections for pulling blocks into the
process of pulling the block. This complicates the locking,
but means that the puller routine has more up-to-date
information. More up-to-date info could lead to faster
pulls, especially as more devices are available.