Naively prefetch block on read
When data is read, prefetch the next block that would be needed
for a sequential read. This naive hueristic relies on the assumption
that most data is read sequentially.

All the appearant complication stems from making locks less
contentious. In the naive implementation, the fmut write lock
would have been held while prefetching. This means, no meaningful
reads on blocks we already have can be executed.

Now, the block puller routines and the reader routines communicate
with a monitor. Multiple reads can now wait on the same block to
be pulled, without any duplication.

This was all driven in the interest of getting video streaming
working. In testing with traffic shaping (not real world), video
watching appears possible if the connection is fast enough and/or
the bit rate is low enough.

This should also be a step that aids file pinning, which was really
how I was planning on tackling video watching.