Improve likelihood of clean shutdown (and thus startup)
For GitHub Issue #11

Sometimes, unmounting wouldn't succeed, and then when SyncthingFUSE
tried to start the next time, the mount point was still in a weird
state, and SyncthingFUSE couldn't mount. I encountered this a few
times, and resolved it by forcibly unmounting when I saw an error.
@fiatjaf also encountered this, and did some digging on it.

With some more research, I was usually able to reproduce the issue
by reading a large file and interrupting SyncthingFUSE in the
middle of the read.

Now, SyncthingFUSE waits a little bit longer for the unmount
request to succeed. I think this primarily gives a little more time
for some socket communication to occur. I think that in between
some of that socket action, the unmount command goes through. The
primary fix here is just waiting. Yay probable race conditions.

I also added a call to stop the service tree. This is a nice fairy
tale, since most services I looked at don't do anything in the
Stop() method. Maybe someday they will, and then this code will
be more useful.